@@ -30,14 +30,14 @@
         }
     ],
     "require": {
-        "php": ">=7.1.3",
+        "php": ">=7.4",
         "lstrojny/functional-php": "~0.1|~1.0",
         "internations/solr-utils": "~0.8"
     },
     "require-dev": {
         "internations/kodierungsregelwerksammlung": "~0.35",
-        "phpunit/phpunit": "~7 || ~8 || ~9",
-        "roave/backward-compatibility-check": "^3 || ^4 || ^5 || ^6"
+        "phpunit/phpunit": "~9",
+        "roave/backward-compatibility-check": "^5 || ^6"
     },
     "autoload": {
         "psr-0": {
